% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AeMFA-data.R
\docType{data}
\name{ADI}
\alias{ADI}
\title{Quarterly revenues of Adidas}
\format{
A data frame with 52 obversations of 2 variables:
\itemize{
\item \code{Datum} date, in the format \code{"YYYY-MM-DD"}
\item \code{Umsatz} revenue in thousand Euros
}
}
\source{
The data were taken from the investor relations web page of Adidas.
}
\usage{
ADI
}
\description{
A dataset containing the quarterly revenues in thousand Euros of the German company
Adidas from 2006 to 2018.
}
\keyword{datasets}
