% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AeMFA.R
\name{AeMFA}
\alias{AeMFA}
\title{AeMFA: Functions of the book \emph{Angewandte empirische Methoden in Finance & Accounting}}
\description{
This package provides the functions which where developed for the book \emph{"Angewandte empirische Methoden in Finance & Accounting"}.
}
