% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AeMFA-data.R
\docType{data}
\name{BankKurse}
\alias{BankKurse}
\title{Daily prices of several European banks}
\format{
A zoo time series with 259 observations of 65 variables:
\itemize{
\item \code{BARC.L} daily price of Barclays bank
\item \code{...} daily price of ..., the name is the respective Yahoo ticker symbol
}
}
\source{
The data were retrieved from Yahoo finance via \code{get_prices_from_tickers()} from
the \code{estudy2} package which uses \code{getSymbols()}
from the \code{quantmod} package.
}
\usage{
BankKurse
}
\description{
A dataset containing the daily prices from several UK and non-UK banks from
2015-07-01 to 2016-06-30.
The prices are given in the respective trading currencies.
}
\keyword{datasets}
