% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AeMFA-data.R
\docType{data}
\name{BEI_1718}
\alias{BEI_1718}
\title{Daily prices of Beiersdorf}
\format{
A xts time series with 507 observations of 1 variable:
\itemize{
\item \code{BEI.DE.Adjusted} daily price in Euro of Beiersdorf
}
}
\source{
The data were retrieved from Yahoo finance via \code{getSymbols()}
from the \code{quantmod} package.
}
\usage{
BEI_1718
}
\description{
A dataset containing the daily (adjusted closing) prices of the German company Beiersdorf from
2016-17-01 to 2017-12-31.
}
\keyword{datasets}
