% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AeMFA-data.R
\docType{data}
\name{CG}
\alias{CG}
\title{Various data in relation to corporate governacne}
\format{
A data frame with 2379 observations of 12 variables:
\itemize{
\item \code{company} name of the company
\item \code{year} year
\item \code{tq} Tobin's Q
\item \code{ext} external corporate governance quality
\item \code{int} internal corporate governance quality
\item \code{lnSize} log of total capital
\item \code{debt} quotient between outside and total capital
\item \code{roa} return on assets, quotient between gross earnings and total capital
\item \code{capex} investments, quotient between investments and total capital
\item \code{cash} quotient between cash plus short term assets and total capital
\item \code{lnResvol} log of residual volatility
\item \code{intangibles} quotient between intangible assets and total capital
}
}
\source{
The data were retrieved from Refinitv Eikon.
}
\usage{
CG
}
\description{
A dataset containing yearly data from 2012 to 2017 about market value,
external and internal corporate governance indicators, and several control
variables of 427 companies listed in the EURO STOXX 600 .
}
\keyword{datasets}
