% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AeMFA.R
\name{CRTestAR}
\alias{CRTestAR}
\title{Corrado rank test for abnormal returns in single company event studies}
\usage{
CRTestAR(ar, wsstart, wsend, westart, weend, nParam = 2)
}
\arguments{
\item{ar}{An \code{xts}-object containing the abnormal returns.}

\item{wsstart}{Date, formated "YYYY-MM-DD", indicating the start of the estimation window.}

\item{wsend}{Date, formated "YYYY-MM-DD", indicating the end of the estimation window.}

\item{westart}{Date, formated "YYYY-MM-DD", indicating the start of the event window.}

\item{weend}{Date, formated "YYYY-MM-DD", indicating the end of the event window.}

\item{nParam}{Number of parameters in the model used for estimating the normal return, e.g.,
2 for the market model and other one-factor models. Defaults to 2.}
}
\value{
A list object containing dates, abnormal returns, abnormal rank values,
standard errors, t-values, and p-values,
plus the cumulative abnormal return, cumulative abnormal rank place, standard error,
t-value, and p-value thereof.
}
\description{
\code{CRTestAR()} applies a Corrado rank test for abnormal returns (Corrado, 1989).
}
\examples{
  AR <- getAR(AeMFA:::BankRet, AeMFA:::IndexRet, "2015-01-01", "2016-06-19")
  CRTestAR(AR, "2015-01-01", "2016-06-19", "2016-06-20", "2016-06-30", 2)

}
\references{
Corrado, C. J. (1989) A nonparametric test for abnormal security-price performance in
event studies, \emph{Journal of Financial Economics}, \bold{23}, 385-–395.
}
\seealso{
\code{getAR()}, \code{tTestAR()}
}
