% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AeMFA-data.R
\docType{data}
\name{CSdf}
\alias{CSdf}
\title{Credit spreads from several countries}
\format{
A data frame with 48 observations of 6 variables:
\itemize{
\item \code{land} country
\item \code{datum} date, formatted as "YYYY-MM-DD"
\item \code{zspread} z-spread of the covered bonds in base points
\item \code{tickets} daily trading tickets
\item \code{restlz} remaining term in days
\item \code{nomzins} nominal interest rate
}
}
\source{
The data were retrieved from Bloomberg.
}
\usage{
CSdf
}
\description{
A dataset containing quarterly data from 2012-03-30 to 2015-12-31 including credit spreads
of covered bonds and various determinants thereof.
}
\keyword{datasets}
