% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AeMFA-data.R
\docType{data}
\name{DAX_1718}
\alias{DAX_1718}
\title{Daily prices of the DAX 30}
\format{
A xts time series with 507 observations of 1 variable:
\itemize{
\item \code{GDAXI.Adjusted} daily price in Euro of the DAX 30
}
}
\source{
The data were retrieved from Yahoo finance via \code{getSymbols()}
from the \code{quantmod} package.
}
\usage{
DAX_1718
}
\description{
A dataset containing the daily (adjusted closing) prices of the German DAX 30 index from
2016-17-01 to 2017-12-31.
}
\keyword{datasets}
