% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AeMFA-data.R
\docType{data}
\name{DIV}
\alias{DIV}
\title{Various determinants in regard to distribution of dividends}
\format{
A data frame with 1281 observations of 20 variables:
\itemize{
\item \code{Name} name of the company
\item \code{Jahr} year
\item \code{Sektor} industrial sector classification according to CIGS
(Consumer, Healthcare, Industrials, IT, Sonstige)
\item \code{Payout} distribution of dividend (0 -- no, 1 -- yes)
\item \code{PayoutVJ} distribution of dividend in the previous year (0 -- no, 1 -- yes)
\item \code{Quote} quotient between dividends and revenues
\item \code{ROA} return on assets
\item \code{FCFUE} quotient between free cashflow and revenues
\item \code{lnUEWachs} log of revenue growth
\item \code{lnInv} log of quotient between net investments and total capital
\item \code{lnFKQ} log of levereage ratio
\item \code{lnCash} log of quotient between cash and total capital
\item \code{lnGroesse} log of company size proxy
\item \code{Float} proportion of free float
\item \code{ARP} share buyback (0 -- no, 1 -- yes)
\item \code{lnKBV} log of price to book ratio
\item \code{Abgeltung} 1 for the years 2009 and 2010 due to change in taxation, else 0
}
}
\source{
The data were retrieved from Bloomberg.
}
\usage{
DIV
}
\description{
A dataset containing yearly data from 2006 to 2010 about market value,
external and internal corporate governance indicators, and several control
variables of 427 companies listed in the EURO STOXX 600.
}
\keyword{datasets}
