% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AeMFA.R
\name{getAR}
\alias{getAR}
\title{Abnormal returns in single company event studies}
\usage{
getAR(company, index, start, end)
}
\arguments{
\item{company}{An \code{xts}-object containing the company returns.}

\item{index}{An \code{xts}-object containing the market index returns.}

\item{start}{Date, formated "YYYY-MM-DD", for the start of the estimation window.}

\item{end}{Date, formated "YYYY-MM-DD", for the end of the estimation window.}
}
\value{
Upon success the function returns a named \code{xts}-object containing
the abnormal returns for the whole period given by \code{company} or \code{index}.
}
\description{
\code{getAR()} calculates the abnormal return over a standard return calculated by a
market model.
}
\examples{
  getAR(company, index, "2010-01-01", "2015-12-31")

}
\seealso{
\code{tTestAR()}, \code{CRTestAR()}
}
