% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AeMFA-data.R
\docType{data}
\name{IndexKurse}
\alias{IndexKurse}
\title{Daily prices of the EURO STOXX 50}
\format{
A zoo time series with 254 observations of 1 variable:
\itemize{
\item \code{^STOXX50E} daily price in Euro of the EURO STOXX 50
}
}
\source{
The data were retrieved from Yahoo finance via \code{get_prices_from_tickers()} from
the \code{estudy2} package which uses \code{getSymbols()}
from the \code{quantmod} package.
}
\usage{
IndexKurse
}
\description{
A dataset containing the daily prices of the EURO STOXX 50 from
2015-07-01 to 2016-06-30.
}
\keyword{datasets}
