% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AeMFA-data.R
\docType{data}
\name{KS}
\alias{KS}
\title{Various determinants for capital structure in companies}
\format{
A data frame with 1375 observations of 19 variables:
\itemize{
\item \code{Name} name of the company
\item \code{Sektor} industrial sector
\item \code{Jahr} year
\item \code{Index} index, where the company is listed (DAX, MDAX, SDAX, TecDAX)
\item \code{fkq} leverage ratio
\item \code{lfk} long term debt (as proportion of total capital)
\item \code{kfk} short term debt (as proportion of total capital)
\item \code{uw} yearly growth of revenues
\item \code{lnGroesse} log of revenues
\item \code{sach} tangible assets (as proportion of total capital)
\item \code{ebit} earning before interest and taxes (as proportion of total capital)
\item \code{ntds} non debt tax shield
\item \code{Steuer} taxes (as proportion of ebit)
\item \code{beta} beta factor
\item \code{BIP} yearly growth of GDP
\item \code{ZS} interest spread between 3-month money market rate and 10-year federal bonds
\item \code{INFL} inflation (yearly change of consumer price index)
\item \code{MP} market performance (yearly return of DAX)
}
}
\source{
The data were retrieved from Refinitiv Eikon.
Macrooeconomical data were retrieved from the OECD and the German federal office
for statistics.
}
\usage{
KS
}
\description{
A dataset containing yearly data from 2005 to 2015 about capital structure and
determinants thereof of 125 German companies.
}
\keyword{datasets}
