% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AeMFA.R
\name{lpAICBIC}
\alias{lpAICBIC}
\title{AIC and BIC for logistic panel regression models}
\usage{
lpAICBIC(mod)
}
\arguments{
\item{mod}{A logistic panel regression model created by \code{bife()}, \code{clogit()},
\code{glmer()}, or \code{pglm()}.}
}
\value{
Upon success the function returns a named numeric vector containing
AIC and BIC (if feasible).
}
\description{
\code{lpAICBIC()} calculates AIC and eventually BIC for models created with
\code{bife()}, \code{clogit()}, \code{glmer()}, or \code{pglm()}.
}
\examples{
  lpAICBIC(AeMFA:::DIV.clogit)

}
\seealso{
\code{pAICBIC()}
}
