% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AeMFA.R
\name{optCP}
\alias{optCP}
\title{Optimal cutpoint for logistic regression models}
\usage{
optCP(mod, yvar, success)
}
\arguments{
\item{mod}{Logistic regression model created by \code{glm()}.}

\item{yvar}{Name of the \eqn{y}-variable.}

\item{sucess}{Value for success (\code{string} or \code{integer}).}
}
\value{
A named numeric.
}
\description{
\code{optCP()} calculates the optimal cutpoint for logistic regression models using the
approach proposed by Wooldridge (2010,  p. 574).
The selected cutpoint is choosen in a way that \eqn{\sum y = \sum \hat y}.
}
\details{

}
\examples{
  optCP(KRED.glm6, "ausfall", "ja")

\dontrun{
  optCP(mod, "name", 1)
  optCP(mod, "name", "yes")
}

}
\references{
Wooldrige, J. M., 2010, \emph{Econometric analysis of cross section and panel data},
2nd ed.,  Cambridge: MIT Press
}
