% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AeMFA.R
\name{pAICBIC}
\alias{pAICBIC}
\title{AIC, AICc, and BIC for panel regression models}
\usage{
pAICBIC(mod)
}
\arguments{
\item{mod}{A panel regression model created by \code{plm()}.}
}
\value{
Upon success the function returns a named numeric vector containing
AIC, AICc, and BIC.
}
\description{
\code{pAICBIC()} calculates AIC, AICc, and BIC for models created with \code{plm()}.
}
\examples{
  pAICBIC(AeMFA:::CS.fem)

}
\seealso{
\code{\link[=lpAICBIC]{lpAICBIC()}}
}
