% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AeMFA-data.R
\docType{data}
\name{SP}
\alias{SP}
\title{Savings plan}
\format{
A data frame with 100 observations of 2 variables:
\itemize{
\item \code{Zeit} time
\item \code{Ersparnis} capital saved
}
}
\source{
The data were simulated.
}
\usage{
SP
}
\description{
A dataset containing simulated data for a savings plan.
}
\keyword{datasets}
