% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AeMFA.R
\name{tTestAR}
\alias{tTestAR}
\title{t-Test for abnormal returns in single company event studies}
\usage{
tTestAR(ar, wsstart, wsend, westart, weend, nParam = 2, flPlot = FALSE)
}
\arguments{
\item{ar}{An \code{xts}-object containing the abnormal returns.}

\item{wsstart}{Date, formated "YYYY-MM-DD", indicating the start of the estimation window.}

\item{wsend}{Date, formated "YYYY-MM-DD", indicating the end of the estimation window.}

\item{westart}{Date, formated "YYYY-MM-DD", indicating the start of the event window.}

\item{weend}{Date, formated "YYYY-MM-DD", indicating the end of the event window.}

\item{nParam}{Number of parameters in the model used for estimating the normal return, e.g.,
2 for the market model and other one-factor models. Defaults to 2.}

\item{flPlot}{Logical, indicates whether a plot of the abnormal returns including an interval
of the critical values of \eqn{0\pm 2} standard errors should be drawn. Defaults to FALSE.}
}
\value{
A list object containing dates, abnormal returns, standard errors, t-values, and p-values,
plus the cumulative abnormal return, standard error, t-value, and p-value thereof.
}
\description{
\code{tTestAR()} applies a t-test for abnormal returns.
}
\examples{
  AR <- getAR(AeMFA:::BankRet, AeMFA:::IndexRet, "2015-01-01", "2016-06-19")
  tTestAR(AR, "2015-01-01", "2016-06-19", "2016-06-20", "2016-06-30", 2, TRUE)

}
\seealso{
\code{\link[=getAR]{getAR()}}, \code{\link[=CRTestAR]{CRTestAR()}}
}
