% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AeMFA-data.R
\docType{data}
\name{UR}
\alias{UR}
\title{Running yield}
\format{
A data frame with 585 observations of 3 variables:
\itemize{
\item \code{Datum} date (formatted \code{"YYYY-MM-DD"})
\item \code{UmlR} running yield in percent
\item \code{InflR} inflation rate in percent
}
}
\source{
The data were retrieved from the web pages of the Federal Bank of Germany.
}
\usage{
UR
}
\description{
A dataset containing monthly data from 1970 to 2018 about running yield and inflation rate.
}
\keyword{datasets}
