% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AeMFA.R
\name{whtest}
\alias{whtest}
\title{Performs a Wu-Hausmann-Test}
\usage{
whtest(mod1, mod2)
}
\arguments{
\item{mod1}{A regression model, tested with \code{clogit()}, \code{pglm()}, \code{plm()},
and \code{iv\_robust()}.}

\item{mod2}{Another model of the classed given above.}
}
\value{
None. A message for the alternative and the values for
  \eqn{\chi^2}, degrees of freedom applied, and the p-value are printed.
}
\description{
\code{whtest()} applies a Wu-Hausmann-Test (Wu, 1973; Hausman, 1978) for various model types.
  \emph{Note:} Not implemented for \code{bife}- and \code{glmer}-models.
  Under the null both models are consistent, under the alternative one of the models
  is inconsistent.
  In this case, a message is displayed which model is inconsistent.
}
\examples{
  whtest(mod1, mod2)

}
\references{
Wu, D.-M. (1973) Alternative tests of independence between stochastic regressors and
  disturbances, \emph{Econometrica} \bold{41}, 733–-750, DOI 10.2307/1914093.

  Hausman, J. A. (1978) Specification tests in econometrics, \emph{Econometrica} \bold{46},
  1251-–1271.
}
