% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AeMFA.R
\name{optBT}
\alias{optBT}
\title{Optimal number of lags for Box.test}
\usage{
optBT(x, maxLags = 20, type = "Box-Pierce", fitdf = 0, flPrint = FALSE)
}
\arguments{
\item{type}{Test to be performed ("Box-Pierce", "Ljung-Box"), partial matching is used.}

\item{fitdf}{Integer, number of degrees of freedom to be subtracted if \code{x} is a series of residuals.}

\item{flPrint}{If TRUE, the tested number of lags and p-values are printed. Defaults to FALSE}

\item{X}{a numeric vector or univariate time series.}

\item{maxlags}{Integer, gives the maximal number of lags to which the test should be performed.}
}
\value{
A list with class "htest" containing the following components:
\itemize{
\item \code{statistic} the value of the test statistic.
\item \code{parameter} the degrees of freedom of the approximate chi-squared distribution of
the test statistic (taking fitdf into account).
\item \code{p.value} the p-value of the test.
\item \code{method}	a character string indicating which type of test was performed.
\item \code{data.name} character string giving the name of the data.
}
}
\description{
\code{optBT()} applies a \code{Box.test()}.
The number of lags were selected according to the smallest p-value within the given range.
}
\details{
Details see the documentation for \code{Box.test()}.
}
\examples{
  x <- rnorm(100)
  optBT(x, maxLags = 30)

}
\seealso{
\code{\link[=Box.test]{Box.test()}}
}
